slowrx: Move listen thread code to listen module
