+++ listen.h
@@ -1,0 +1,6 @@
+#ifndef _LISTEN_H_
+#define _LISTEN_H_
+
+void     *Listen       ();
+
+#endif
